pyproject.toml
[build-system]
requires = ["setuptools==75.8.0", "wheel==0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "greenova"
version = "1.0.0"
description = "Environmental Compliance Management System"
readme = "README.md"
requires-python = "==3.13.1"
license = {text = "AGPL-3.0"}
authors = [
    {name = "Adrian Gallo", email = "agallo@enveng-group.com.au"},
]
dependencies = [
    "pip==25.0",
    "asgiref==3.8.1",
    "cffi==1.17.1",
    "cryptography==41.0.7",
    "Django==5.1.5",
    "django-appconf==1.0.6",
    "pillow==11.1.0",
    "pycparser==2.22",
    "rcssmin==1.1.2",
    "rjsmin==1.2.2",
    "setuptools==75.8.0",
    "sqlparse==0.5.3",
    "wheel==0.45.1",
    "python-dotenv==1.0.1",
    "django-compressor==4.5.1",
    "whitenoise==6.8.2",
]

[project.optional-dependencies]
dev = [
    "editorconfig==0.17.0",
    "autopep8==2.3.2",
    "debugpy==1.8.2",
]

[project.urls]
Homepage = "https://github.com/enssol/greenova"
Repository = "https://github.com/enssol/greenova.git"
Documentation = "https://greenova.readthedocs.io/"

[tool.setuptools]
packages = ["greenova"]

[tool.unittest]
start-directory = "tests"
pattern = "test_*.py"
verbosity = 2

[tool.autopep8]
max_line_length = 79
in-place = true
recursive = true
